[PATCH] ppc64: make HvLpEvent_unregisterHandler() work
When the iseries_veth driver module is unloaded there is the potential for
an oops and also some memory leakage.

Because the HvLpEvent_unregisterHandler() function did no synchronisation,
it was possible for the handler that was being unregistered to be running
on another CPU *after* HvLpEvent_unregisterHandler() had returned.  This
could cause the iseries_veth driver to leave work in the events work queue
after the module had been unloaded.  When that work was eventually executed
we got an oops.

In addition some of the data structures in the iseries_veth driver were not
being correctly freed when the module was unloaded.

This is the first patch, which makes HvLpEvent_unregisterHandler() work.

Signed-off-by: Michael Ellerman <michael@ellerman.id.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e9a577pBEWRz9_SDKe56_1bPAoHg