[PATCH] ppc64: iommu: avoid ISA io space on POWER3
On some systems, the first PCI bus has a ISA I/O hole at the first 16MB.
We can't use this space for DMA addresses on the bus.

On Python-based machines, we'll skip the first 256MB on buses that have the
hole, just as we do on later systems.  This means that the first bus will
have 768MB of DMA space shared between the devices on it.

Signed-off-by: Olof Johansson <olof@austin.ibm.com>
Acked-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e9a596Iift_ZeOW5_b0s0eBMDMQw