[PATCH] uml: Makefile simplification and correction.
Cleanup: simplify a lot of strange constructs and whatever present in
arch/um/Makefile.

Also, get rid of redundant cleaning code introduced in
"uml-fix-make-clean.patch" from 2.6.10-mm3 - when it was written it made
sense, but I fixed most problems it addressed in a more elegant way.

Also about that, don't remove $(ARCH_SYMLINKS) in make clean, but rather in
make mrproper as we already do, like for include/asm-um and other symlinks.

Finally, remove one wrong thing (almost a bug) introduced in that - the usage
of the clean-dirs construct:

clean-dirs := sys-$(SUBARCH)

which is intended to delete one whole folder, rather than to descend to clean
it, when used in normal Makefiles (not in the arch Makefile where is used,
with no effect).  It's also not needed because that folder is cleaned because
is listed in either $(core-y) or $(libs-y).

Signed-off-by: Paolo 'Blaisorblade' Giarrusso <blaisorblade_spam@yahoo.it>
Cc: Chris Wright <chrisw@osdl.org>
Cc: Jeff Dike <jdike@addtoit.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e9a72fvrVZQRnPLYeQ7wuc14et_Q