[PATCH] pcmcia: ignore driver IRQ mask
As its just one pin which can be used for IRQs on PCMCIA/CardBus cards, and
only the _socket drivers_ care with which IRQ this pin is connected, only the
socket drivers (assisted by the PCCARD core) should care about which IRQ to
use.  Therefore, ignore the information passed to pcmcia_request_irq() in
IRQInfo2.  In additional patches, all in-kernel users of IRQInfo2 will be
removed; users wishing to influence the usage of IRQs can do so by modifying
several driver's irq_mask parameter and/or by adding "exclude irq 3" to
/etc/pcmcia/config.opts.  Note that a new sysfs-based interface to do so will
be added in subsequent patches.

Signed-off-by: Dominik Brodowski <linux@brodo.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e9a9daIQ7lTQeTtKrt32GXSHT0NQ