[PATCH] ioctl compatibility for TIOCMIWAIT and TIOCGICOUNT
This patch lets us use TIOCMIWAIT and TIOCGICOUNT from a 32-bit process on
a 64-bit processor.  TIOCMIWAIT uses the argument as a bitmap of things to
wait for.  The argument for TIOCGICOUNT points to a struct
serial_icounter_struct, which only contains ints and arrays of int.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41f124ddJNHwL4NLHo2DztKYq6AE5w