[PATCH] ppc32: System platform_device description, discovery and management
On most embedded PPC systems we either have a core CPU and chipset (MPC10x,
TSI10x, Marvell, etc.) or a system-on-chip device (4xx, 8xx, 82xx, 85xx,
etc.).  Some of these sub-archs have been using the On Chip Peripheral
(OCP) driver model.  The functionality that OCP provide has been replaced
by the generic driver model and platform_device.  Also, some of these
device may exist across a number of architectures (PPC, MIPS, ARM) such
that some information that is shared between the architecture and driver
needs to exist outside of either.

The ppc_sys changes add a standard way for PowerPC systems to describe the
devices and systems that exist in the sub-arch.  Additionally, we are able
to discover which system we are and manage which devices are actually
registered and any platform specific fixups that may be needed.

Signed-off-by: Kumar Gala <kumar.gala@freescale.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41f12533Y-lASqJxX9xsA6Awvf-_6g