[PATCH] move read-only and immutable checks into permission()
Currently it's up to the filesystem ->permission method to check whether
the filesystem is readonly or the file marked immutable.  But this is
really a VFS decision, and the distintion becomes more important when
moving to per-mountpoint read only flags.

For most filesystems that have been using generic_permission this is not
change in behaviour.  For those that we're missing the check (cifs without
CIFS_MOUNT_NO_PERM and coda [1]) this is a bugfix.

Both reiserfs and xfs have this check still in their ->permission routine
because they call it from other places aswell.  I'll try switching them
over to generic_permission and will take care of this one.

[1] coda_ioctl_permission always returns 0, aka always grants access,
    which looks more than fishy to me.


Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41f128029a9x3ihQ5RNzThtVcV6nDQ