[PATCH] ext3: commit superblock before panicking
I have a problem with errors=panic on ext3.  When a panic occurs, the error
event is not recorded anywhere.  So after the reboot, e2fsck doesn't kick
in, the file system gets mounted again and the box panics again...

Patch below moves the ERRORS_PANIC test down a bit so the journal is
aborted before panic() is called.

Signed-off-by: Eric Lammerts <eric@lammerts.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41f128adfD0zQ7mv0-dFY5Ne-GjAvg