[PATCH] add page_offset to mm.h
To convert page->index to a byte index you need to cast it to loff_t first
so it's a 64bit value.  There have been quite a few places that got it
wrong in the kernel.  To make it easier a nice little helper would be nice,
and in fact the NFS code already has it.  Let's move it to pagemap.h so
everyone can use it.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41f128eehU_8kdPqeKo-uCMB1ExwTQ