[IA64] uaccess.h: add missing __user annotation for sparse
I was getting a lot of spurious warnings with "make C=1", due to a
missing "__user" attribute.  I'm not sure whether this got lost or
whether earlier versions of sparse failed to warn about it, but it's
clearly needed (and the i386 version of uaccess.h does the same).

Signed-off-by: David Mosberger-Tang <davidm@hpl.hp.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

BKrev: 41f2bba35yaf10v_YHbsOvHmUH_zXA