[IA64] Use srlz.d instead of srlz.i in ia64_leave_syscall
Use srlz.d instead of srlz.i.  Safe because we don't care whether
the VHPT walker sees the clearing of PSR.ic (if it does, that's fine
if it doesn't, it's OK too since the kernel-text is pinned anyhow).
Good for another 11+ cycles in (normal) getpid().

Signed-off-by: David Mosberger-Tang <davidm@hpl.hp.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

BKrev: 41f2c6adZCdvhmjXcN1n_r98UuXekQ