[PATCH] removing bcopy... because it's half broken
Nothing in the kernel is using bcopy right know, and that is a good thing.
Why? Because a lot of the architectures implement a broken bcopy()....
the userspace standard bcopy() is basically a memmove() with a weird
parameter order, however a bunch of architectures implement a memcpy() not a
memmove().

Instead of fixing this inconsistency, I decided to remove it entirely,
explicit memcpy() and memmove() are prefered anyway (welcome to the 1990's)
and nothing in the kernel is using these functions, so this saves code size
as well for everyone.

Signed-off-by: Arjan van de Ven <arjan@infradead.org>
Signed-off-by: Adrian Bunk <bunk@stusta.de>

[ Side note: the only reason for bcopy appears to be totally ancient
  gcc versions for OSF/1, used to originally cross-compile Linux on
  alpha. Possibly some other similar cases. Time to move on ;-]

Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41f58b60X1DD4y9ru-z53GUStLSUcQ