[PATCH] ppc64: clear MSR_RI earlier in syscall exit path
This patch is from Craig Chaney <cchaney@us.ibm.com>.

This patch moves the restoring of the stack pointer in the system call exit
path to after the point where we clear the RI (recoverable interrupt) bit in
the MSR.  Normally, loading the stack pointer before clearing RI doesn't cause
any problem because there is no trap that can normally occur in between.  But
if we are tracing the code using a tool that single-steps instructions, this
can cause a problem.  In this case, clearing RI serves as an indication that
the following code can't be safely single-stepped.

Signed-off-by: Craig Chaney <cchaney@us.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41f71dc3aLVjpHc4IuuxZS-N7xphmA