[PATCH] ext2/ext3 ACLs: remove the number of acl entries limit
This patch removes the arbitrary limit of 32 acl entries on ext[23] when
writing acls.  A patch that removes the same check when reding acls is in
BK since 12 March 2004, so all kernels since then are already able to read
large acls.  I think that ten+ months are enough so that we can now also
remove the write limit.

This is the read-limit patch:
http://linux.bkbits.net:8080/linux-2.6/cset%404051e2863UsuQEgAQShmimgBooAXkg?nav=index.html

Even without this patch the xattr block could already contain less space
than needed for the acl, because other attributes might already use up
almost all space.  So this patch does not introduce additional error
conditions.  We have been shipping with this patch the last year (almost).

Signed-off-by: Andreas Gruenbacher <agruen@suse.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41f71fc9lVUWJhYy0FPDn0SfAfyfkw