[PATCH] fix broken cross compiles
The change to kbuild to use -isystem `gcc -print-file-name=include`
broke our parisc crosscompile (and presumably everyone else's).

The reason is that you have a := in the NOSTDINC_FLAGS rule, which is
evaluated in situ (i.e.  before we've had a chance to set CROSSCOMPILE
on CC) so the gcc include path is actually the native one not the
crosscompiler one.  On parisc this causes us to be unable to handle
_builtin_va functions, but I bet there are a heap of other problems.

Acked-by: Sam Ravnborg <sam@ravnborg.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41f731f7Di9jQdUixEetQ4FiKp4uoQ