[PATCH] random: overflow fix
Zwane reports an oops duie to add_entropy_words running off the top of the
stack reading its input.  The overrun is harmless until it causes a page
fault.

Looks like I introduced this bug in Aug 2003, but it was hard to trigger until
the recent changes.

Signed-off-by: Matt Mackall <mpm@selenic.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4201000756W9mfU5Sm0aBEnattI_6Q