[PATCH] alloc_zeroed_user_highpage() to fix the clear_user_highpage issue
This patch adds a new function alloc_zeroed_user_highpage that is then used
in the anonymous page fault handler and in the COW code to allocate zeroed
pages.  The function can be defined per arch to setup special processing
for user pages by defining __HAVE_ARCH_ALLOC_ZEROED_USER_PAGE.  For arches
that do not need to do special things for user pages,
alloc_zeroed_user_highpage is defined to simply do

	alloc_page_vma(GFP_HIGHUSER | __GFP_ZERO, vma, vaddr)

This patch needs to update a number of archs. Wish there was a better way
to do this.

Signed-off-by: Christoph Lameter <clameter@sgi.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 42010109JqOPNWQBb20y6ATvnrF3IQ