[PATCH] ppc32: pmac sleep support update
This patch updates the PowerMac sleep support.  The ability to sleep is now
broken into 2 different flags, one, "may sleep" is set for all motherboards
that we know how to put to sleep and wakeup.  It gets turned into "can sleep"
upon a call from the video driver indicating the ability to wakeup the video
card.  This doesn't deal with head-less machines, but this can be improved
later.  It also adds better cache flush code, which improves stability with
cpufreq as well as sleep.

This patch actually breaks sleep support until the video drivers for the
affected machines have been updated.  This will come as separate patches.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 42010137eNuC9Fwegxk1MrsprC9BEg