[PATCH] ppc32: Fix via IDE driver for pegasos
The Pegasos machines have the VIA IDE controller configured in a strange
way by the firmware.  It claims to be fully native, but is actually routed
to legacy irqs 14 and 15.

This patch works around that in the driver.  In the future, I'll try to do
an ifdef-less patch by detecting that setup from the chip, and using the
new pci_get_legacy_ide_irq() in that case, but I don't have proper HW
infos.  This patch is good enough in the meantime, I'll come up with a
better one later eventually.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Acked-by: Bartlomiej Zolnierkiewicz <bzolnier@elka.pw.edu.pl>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 42010191utbssmut6j1X9p7FDaiILQ