[PATCH] avma1_cs: Inverted parameter order in outb
avma1_cs detect routine for AVM A1 based PCMCIA cards has its "outb"
parameters inverted - switch to "byteout" define to be conformant with the
rest of the driver, fixing the order.

This was noticed on PPC, which triggers an MCE in case of invalid IO port
access, which is not the case on x86.

Signed-off-by: Marcelo Tosatti <marcelo.tosatti@cyclades.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4202c952VIX1nQVo-PikWKxNaRlJ8g